Moved meshing to separated mod
