+++ crates/world_server/src/set/mod.rs
@@ -1,9 +1,11 @@
 mod chunk_initialization;
 mod chunk_management;
 mod landscape;
+mod meshing;
 mod propagation;
 
 pub use chunk_initialization::*;
 pub use chunk_management::*;
 pub use landscape::*;
+pub use meshing::*;
 pub use propagation::*;
